\name{P_efa}
\alias{P_efa}
\docType{data}
\title{
The correct permutation matrix, P_i, for each replication
}
\description{
A list containing 200 correct P_i to re-order and/or re-sign as needed within the relevant parameter estimates provided by replication numbers 1 through 100 and 4701 through 4800 in Example 1 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(P_efa)}
\format{
This dataset contains a list containing 200 correct 3*3 P_i to re-order and/or re-sign as needed within the relevant parameter estimates provided by replication numbers 1 through 100 and 4701 through 4800 in Example 1 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
