\name{P_esem}
\alias{P_esem}
\docType{data}
\title{
The correct permutation matrix, P_i, for each replication             
}
\description{
A list containing 200 correct P_i to re-order and/or re-sign as needed within the relevant parameter estimates provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(P_esem)}
\format{
This dataset contains a list containing 200 correct 4*4 P_i to re-order and/or re-sign as needed within the relevant parameter estimates provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
