\name{pop_L_efa}
\alias{pop_L_efa}
\docType{data}
\title{
The population pattern coefficient matrix
}
\description{
The population pattern coefficient matrix in Example 1 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(pop_L_efa)}
\format{
 This dataset contains the population 10*3 pattern coefficient matrix in Example 1 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
