\name{pop_L_esem}
\alias{pop_L_esem}
\docType{data}
\title{
The population pattern coefficient matrix             
}
\description{
The population pattern coefficient matrix in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(pop_L_esem)}
\format{
 This text file contains the population 11*4 pattern coefficient matrix in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).  
}
