\name{sample_alpha_esem}
\alias{sample_alpha_esem}
\docType{data}
\title{
Replications of the estimated intercept or mean vector
}
\description{
A list containing 200 replications of the estimated intercept or mean vector provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(sample_alpha_esem)}
\format{
This dataset is a list that contains 200 replications of the estimated 4*1 intercept or mean vector provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
