\name{sample_beta_esem}
\alias{sample_beta_esem}
\docType{data}
\title{
Replications of the estimated eta on eta regression coefficient matrix
}
\description{
A list containing 200 replications of the estimated eta on eta regression coefficient matrix provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(sample_beta_esem)}
\format{
This dataset is a list that contains 200 replications of the estimated 4*4 eta on eta regression coefficient matrix provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
