\name{sample_gamma_esem}
\alias{sample_gamma_esem}
\docType{data}
\title{
Replications of the estimated eta on x regression coefficient matrix
}
\description{
A list containing 200 replications of the estimated eta on x regression coefficient matrix provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(sample_gamma_esem)}
\format{
This datset is a list that contains 200 replications of the estimated 4*1 eta on x regression coefficient matrix provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
