\name{sample_lambda_efa}
\alias{sample_lambda_efa}
\docType{data}
\title{
Replications of the estimated pattern coefficient matrix
}
\description{
A list containing 200 replications of the estimated pattern coefficient matrix provided by replication numbers 1 through 100 and 4701 through 4800 in Example 1 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(sample_lambda_efa)}
\format{
This dataset is a list that contains 200 replications of the estimated 10*3 pattern coefficient matrix provided by replication numbers 1 through 100 and 4701 through 4800 in Example 1 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
