\name{sample_lambda_esem}
\alias{sample_lambda_esem}
\docType{data}
\title{
Replications of the estimated pattern coefficient matrix
}
\description{
A list containing 200 replications of the estimated pattern coefficient matrix provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(sample_lambda_esem)}
\format{
This dataset is a list that contains 200 replications of the estimated 11*4 pattern coefficient matrix provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
