\name{sample_psi_esem}
\alias{sample_psi_esem}
\docType{data}
\title{
Replications of the estimated covariance matrix for the vector of residuals for eta
}
\description{
A list containing 200 replications of the estimated covariance matrix for the vector of residuals for eta provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
\usage{data(sample_psi_esem)}
\format{
This dataset is a list that contains 200 replications of the estimated 4*4 covariance matrix for the vector of residuals for eta provided by replication numbers 1 through 100 and 4701 through 4800 in Example 2 from Myers, Ahn, Lu, Celimli, and Zopluoglu (2016).
}
